requirements.txt
selenium==4.17.2
pytest==7.4.3
pytest-html==4.1.1
webdriver-manager==4.0.1
Pillow>=11.0.0
